Simple alternative to https://song.link
Create a preview of a YT video. This can be distributed in the Fediverse. In addition, the application can be used as an alternative to fat Google search. 
A google api-key (YouTube Data API v3 ) https://console.cloud.google.com is required for use. The api key must be stored in config.php.